fix: change header on scroll
