+++ src/Components/App/App.module.css
@@ -23,6 +23,6 @@
   padding: 1rem;
 }
 
-.title-header{
+.title-header {
   text-align: center;
 }
