+++ src/Components/MUIHeader/MUIHeader.types.ts
@@ -1,0 +1,10 @@
+export interface ChangeOnScrollProps {
+  children: React.ReactElement;
+}
+
+export type MUIHeaderProps = {
+  pages: string[];
+  isLogin: boolean;
+  lang: boolean;
+  setLang: (c: boolean) => void;
+};
