+++ src/pages/EditorPage/style.module.scss
@@ -15,6 +15,5 @@
 .test-div {
   height: 800px;
   width: 100vw;
-  // background-color: #f5f5dc;
   background-color: #a0a0a081;
 }
